move all menu logic to MobileMenu component
