+++ src/components/layouts/MobileMenu.tsx
@@ -1,17 +1,22 @@
-import { useEffect, useRef } from 'react';
-import { Box, Flex, FlexProps, Input, InputGroup, Link, Stack, Text } from '@chakra-ui/react';
+import { useState, useEffect, useRef } from 'react';
+import { Box, Flex, Input, InputGroup, Link, Stack, Text } from '@chakra-ui/react';
+import { CloseIcon } from '@chakra-ui/icons';
 import NextLink from 'next/link';
 
-import { CloseIcon } from '@chakra-ui/icons';
-import { LensIcon } from '../UI/icons';
+import { HamburgerIcon, LensIcon } from '../UI/icons';
 import { DOCS_PAGE, DOWNLOADS_PAGE } from '../../constants';
 
-interface Props extends FlexProps {
-  setIsMenuOpen: (value: boolean) => void;
-}
-export const MobileMenu: React.FC<Props> = ({ setIsMenuOpen, ...props }) => {
+const BORDER_WIDTH = '2px';
+
+export const MobileMenu: React.FC = () => {
+  const [isMenuOpen, setIsMenuOpen] = useState<boolean>(false);
   const containerRef = useRef<HTMLDivElement>(null)
   
+  const handleMenuToggle = () => {
+    setIsMenuOpen(prev => !prev);
+  }
+
+
   // Event listening for keydown or click
   useEffect(() => {
     const handleKeyPress = (event: KeyboardEvent): void => {
@@ -34,9 +39,10 @@ export const MobileMenu: React.FC<Props> = ({ setIsMenuOpen, ...props }) => {
   
   const mobileMenuItemStyles = {
     justifyContent: 'center',
-    borderBottom: '2px',
+    borderBottom: BORDER_WIDTH,
     borderColor: 'bg',
     color: 'primary',
+    p: 8,
     _hover: {
       textDecoration: 'none',
       bg: 'primary',
@@ -45,75 +51,95 @@ export const MobileMenu: React.FC<Props> = ({ setIsMenuOpen, ...props }) => {
   }
 
   return (
-    <Flex
-      position='absolute'
-      top="-2px"
-      insetInline="-2px"
-      color='bg'
-      bg='secondary'
-      border='2px'
-      borderColor='bg'
-      ref={containerRef}
-      zIndex={99}
-      transition='all 0.25s ease-in-out'
-      overflow='hidden'
-      direction='column'
-      {...props}
-    >
-      {/* HAMBURGER MENU */}
-      <Flex borderBottom='2px' borderColor='bg' justify='flex-end'>
-        <Box
-          as='button'
-          p={4}
-          color='primary'
-          _hover={{ bg: 'secondary', color: 'bg' }}
-          onClick={() => setIsMenuOpen(false)}
-          ms='auto'
-        >
-          <CloseIcon boxSize={5} color='bg' _hover={{ color: 'primary' }} />
-        </Box>
+    <>
+      {/* HAMBURGER MENU ICON */}
+      <Box
+        as='button'
+        p={4}
+        display={{ base: 'block', md: 'none' }}
+        color='primary'
+        _hover={{
+          bg: isMenuOpen ? 'secondary' : 'primary',
+          color: 'bg'
+        }}
+        onClick={handleMenuToggle}
+        ml={isMenuOpen ? 'auto' : 'none'}
+      >
+        <HamburgerIcon />
+      </Box>
+
+      {/* MOBILE MENU */}
+      <Flex
+        position='absolute'
+        top={`-${BORDER_WIDTH}`}
+        insetInline={`-${BORDER_WIDTH}`}
+        color='bg'
+        bg='secondary'
+        border={BORDER_WIDTH}
+        borderColor='bg'
+        ref={containerRef}
+        zIndex={99}
+        overflow='hidden'
+        direction='column'
+        transform={`translateY(${!isMenuOpen ? '-150%' : '0'})`}
+        h='100vh'
+        transition='all 0.25s ease-in-out'
+      >
+        <Flex borderBottom='2px' borderColor='bg' justify='flex-end'>
+          {/* CLOSE ICON */}
+          <Box
+            as='button'
+            p={4}
+            borderLeft={BORDER_WIDTH}
+            borderColor='bg'
+            _hover={{ bg: 'primary' }}
+            onClick={() => setIsMenuOpen(false)}
+            ml='auto'
+          >
+            <CloseIcon boxSize={5} color='bg' />
+          </Box>
+        </Flex>
+
+        <Flex position="relative" direction='column' w='full'/*  sx={{ '&>div': { p: 8 }}} */>
+          {/* DOWNLOADS */}
+          <Stack {...mobileMenuItemStyles}>
+            <NextLink href={DOWNLOADS_PAGE} passHref>
+              <Link _hover={{ textDecoration: 'none' }} onClick={() => setIsMenuOpen(false)}>
+                <Text textStyle='downloads-button-label' fontSize="2xl">
+                  downloads
+                </Text>
+              </Link>
+            </NextLink>
+          </Stack>
+
+          {/* DOCUMENTATION */}
+          <Stack {...mobileMenuItemStyles}>
+            <NextLink href={DOCS_PAGE} passHref>
+              <Link _hover={{ textDecoration: 'none' }} onClick={() => setIsMenuOpen(false)}>
+                <Text textStyle='downloads-button-label' fontSize="2xl">
+                  documentation
+                </Text>
+              </Link>
+            </NextLink>
+          </Stack>
+
+          {/* SEARCH */}
+          <Stack borderBottom={BORDER_WIDTH} borderColor='bg' p={8} _hover={{ bg: 'primary'}}>
+            <InputGroup>
+              <Input
+                variant='unstyled'
+                placeholder='search'
+                size='md'
+                _placeholder={{ color: 'bg', fontStyle: 'italic' }}
+              />
+              <Stack pl={4} justifyContent='center' alignItems='center'>
+                <LensIcon color='bg' fontSize='2xl' />
+              </Stack>
+            </InputGroup>
+          </Stack>
+        </Flex>
+
       </Flex>
-
-      <Flex position="relative" direction='column' w='full' sx={{ '&>div': { p: 8 }}}>
-        {/* DOWNLOADS */}
-        <Stack {...mobileMenuItemStyles}>
-          <NextLink href={DOWNLOADS_PAGE} passHref>
-            <Link _hover={{ textDecoration: 'none' }} onClick={() => setIsMenuOpen(false)}>
-              <Text textStyle='downloads-button-label' fontSize="2xl">
-                downloads
-              </Text>
-            </Link>
-          </NextLink>
-        </Stack>
-
-        {/* DOCUMENTATION */}
-        <Stack {...mobileMenuItemStyles}>
-          <NextLink href={DOCS_PAGE} passHref>
-            <Link _hover={{ textDecoration: 'none' }} onClick={() => setIsMenuOpen(false)}>
-              <Text textStyle='downloads-button-label' fontSize="2xl">
-                documentation
-              </Text>
-            </Link>
-          </NextLink>
-        </Stack>
-
-        {/* SEARCH */}
-        <Stack borderBottom='2px' borderColor='bg' _hover={{ bg: 'primary'}}>
-          <InputGroup>
-            <Input
-              variant='unstyled'
-              placeholder='search'
-              size='md'
-              _placeholder={{ color: 'bg', fontStyle: 'italic' }}
-            />
-            <Stack pl={4} justifyContent='center' alignItems='center'>
-              <LensIcon color='bg' fontSize='2xl' />
-            </Stack>
-          </InputGroup>
-        </Stack>
-
-      </Flex>
-
-    </Flex>
+    </>
   );
 };
